+++ myems-api/core/combinedequipment.py
@@ -382,7 +382,7 @@ class CombinedEquipmentItem:
     @staticmethod
     @user_logger
     def on_post(req, resp, id_):
-        """Handles PUT requests"""
+        """Handles POST requests"""
         access_control(req)
         if not id_.isdigit() or int(id_) <= 0:
             raise falcon.HTTPError(falcon.HTTP_400, title='API.BAD_REQUEST',
@@ -631,7 +631,6 @@ class CombinedEquipmentEquipmentCollection:
         add_row = (" INSERT INTO tbl_combined_equipments_equipments (combined_equipment_id, equipment_id) "
                    " VALUES (%s, %s) ")
         cursor.execute(add_row, (id_, equipment_id,))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()
@@ -1538,7 +1537,6 @@ class CombinedEquipmentMeterCollection:
         add_row = (" INSERT INTO tbl_combined_equipments_meters (combined_equipment_id, meter_id, is_output ) "
                    " VALUES (%s, %s, %s) ")
         cursor.execute(add_row, (id_, meter_id, is_output))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()
@@ -1734,7 +1732,6 @@ class CombinedEquipmentOfflineMeterCollection:
                    " (combined_equipment_id, offline_meter_id, is_output ) "
                    " VALUES (%s, %s, %s) ")
         cursor.execute(add_row, (id_, offline_meter_id, is_output))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()
@@ -1929,7 +1926,6 @@ class CombinedEquipmentVirtualMeterCollection:
                    " (combined_equipment_id, virtual_meter_id, is_output ) "
                    " VALUES (%s, %s, %s) ")
         cursor.execute(add_row, (id_, virtual_meter_id, is_output))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()