+++ myems-api/core/equipment.py
@@ -396,7 +396,7 @@ class EquipmentItem:
     @staticmethod
     @user_logger
     def on_post(req, resp, id_):
-        """Handles PUT requests"""
+        """Handles POST requests"""
         if not id_.isdigit() or int(id_) <= 0:
             raise falcon.HTTPError(falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_EQUIPMENT_ID')
@@ -1347,7 +1347,6 @@ class EquipmentMeterCollection:
         add_row = (" INSERT INTO tbl_equipments_meters (equipment_id, meter_id, is_output ) "
                    " VALUES (%s, %s, %s) ")
         cursor.execute(add_row, (id_, meter_id, is_output))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()
@@ -1538,7 +1537,6 @@ class EquipmentOfflineMeterCollection:
         add_row = (" INSERT INTO tbl_equipments_offline_meters (equipment_id, offline_meter_id, is_output ) "
                    " VALUES (%s, %s, %s) ")
         cursor.execute(add_row, (id_, offline_meter_id, is_output))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()
@@ -1730,7 +1728,6 @@ class EquipmentVirtualMeterCollection:
         add_row = (" INSERT INTO tbl_equipments_virtual_meters (equipment_id, virtual_meter_id, is_output ) "
                    " VALUES (%s, %s, %s) ")
         cursor.execute(add_row, (id_, virtual_meter_id, is_output))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()