+++ myems-api/core/knowledgefile.py
@@ -148,8 +148,6 @@ class KnowledgeFileCollection:
                 cnx.close()
             raise falcon.HTTPError(falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_USER_PLEASE_RE_LOGIN')
-        else:
-            user_id = row[0]
 
         cnx = mysql.connector.connect(**config.myems_system_db)
         cursor = cnx.cursor()