+++ myems-api/core/meter.py
@@ -960,7 +960,6 @@ class MeterPointCollection:
         add_row = (" INSERT INTO tbl_meters_points (meter_id, point_id) "
                    " VALUES (%s, %s) ")
         cursor.execute(add_row, (id_, new_values['data']['point_id'],))
-        new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
         cnx.close()